fixed movie details page reloading issues using ssr
favorites button in movie dynamic route disabled due to conflicts between localstorage and SSR
